[Refactor] 코드 리팩토링 완료
