+++ client/css/main.css
@@ -743,18 +743,14 @@ html {
   width: 100%;
 }
 
-.w-\[19\.2px\] {
-  width: 19.2px;
-}
-
-.w-\[19\.21px\] {
-  width: 19.21px;
-}
-
 .scale-\[1\.3\] {
   --tw-scale-x: 1.3;
   --tw-scale-y: 1.3;
   transform: translate(var(--tw-translate-x), var(--tw-translate-y)) rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y)) scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
+}
+
+.cursor-pointer {
+  cursor: pointer;
 }
 
 .flex-col {
@@ -828,6 +824,10 @@ html {
   row-gap: 14px;
 }
 
+.overflow-hidden {
+  overflow: hidden;
+}
+
 .rounded-\[10px\] {
   border-radius: 10px;
 }
@@ -861,10 +861,6 @@ html {
 
 .bg-\[linear-gradient\(to_bottom\2c rgba\(0\2c 0\2c 0\2c 0\)\2c rgba\(255\2c 255\2c 255\2c 1\)\2c rgba\(255\2c 255\2c 255\2c 1\)\)\2c url\(\'\/assets\/images\/foods\.jpg\'\)\] {
   background-image: linear-gradient(to bottom,rgba(0,0,0,0),rgba(255,255,255,1),rgba(255,255,255,1)),url('/assets/images/foods.jpg');
-}
-
-.bg-\[url\(\'\/assets\/icons\/Icon\/favorite_review_default\.svg\'\)\] {
-  background-image: url('/assets/icons/Icon/favorite_review_default.svg');
 }
 
 .bg-\[url\(\'\/assets\/icons\/Icon\/views_review\.svg\'\)\] {
