+++ client/lib/utils/index.js
@@ -5,3 +5,4 @@ export * from './memo.js';
 export * from './typeOf.js';
 export * from './tiger.js';
 export * from './storage.js';
+export * from './changeImage.js';
